Updated styles for interactiveatoms with csp
